refactor(commands): extract /share and /links commands into share.rs
